Extract a test
